requirements.txt
langchain_core==0.3.76
langchain_experimental==0.3.4
langchain_openai==0.3.33
oci==2.150.1
oracledb==3.3.0
pandas==2.3.2
pypdf==6.1.0
python-dotenv==1.1.1
streamlit==1.44.1
